+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+* #### Language module removed
+
+  The Language module has been removed, use I18n.default_locale
+  instead of Language.default to specify the default locale.
+
+  3 letter language codes have been deprecated in favor of
+  ISO639-1.
+
 * #### JSON rendering for pages
 
   Pages can now be accessed via JSON.